--- Timesheets/Services/CSVExportServicecs.cs
@@ -1,19 +1,0 @@
-﻿using System.Text;
-using Timesheets.Models;
-
-namespace Timesheets.Services
-{
-    public interface ICsvService
-    {
-        string CreateCSV(IList<Timesheet> timesheets);
-    }
-
-    public class CSVExportServicecs : ICsvService
-    {
-        public string CreateCSV(IList<Timesheet> timesheets)
-        {
-       
-            return "";
-        }
-    }
-}
